Add optional HR interpolation
The interpolation is optional, but when not in use, it actually just
interpolates within a max time of 1 second. This means that times no
longer have to exactly match on the second, but instead just have to be
within 1 second of each other. This fixes the issue where GPX files with
fractional seconds never had any data added to them.